backend/requirements.txt
Django==5.2.5
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
Pillow==10.1.0
python-decouple==3.8
